Setup a feedback loop for exploring the API.
The first task was to explore the dev.to API, understand what I can
request in order to create an article, and how they are laid out.

Furthermore, as outgoing HTTP requests needs to go through their own
`requestUrl` method, we needed a wrapper allowing us to express the code
using that interface; while actually making a real call during the
exploration.